add a custom `Writer` for inflate
