+++ zlib-rs/src/read_buf.rs
@@ -197,98 +197,6 @@ impl<'a> ReadBuf<'a> {
         self.filled = end;
     }
 
-    #[inline(always)]
-    pub fn copy_match(&mut self, offset_from_end: usize, length: usize) {
-        #[cfg(target_arch = "x86_64")]
-        if crate::cpu_features::is_enabled_avx512() {
-            return self.copy_match_help::<core::arch::x86_64::__m512i>(offset_from_end, length);
-        }
-
-        #[cfg(target_arch = "x86_64")]
-        if crate::cpu_features::is_enabled_avx2() {
-            return self.copy_match_help::<core::arch::x86_64::__m256i>(offset_from_end, length);
-        }
-
-        #[cfg(target_arch = "x86_64")]
-        if crate::cpu_features::is_enabled_sse() {
-            return self.copy_match_help::<core::arch::x86_64::__m128i>(offset_from_end, length);
-        }
-
-        self.copy_match_help::<u64>(offset_from_end, length)
-    }
-
-    #[inline(always)]
-    fn copy_match_help<C: Chunk>(&mut self, offset_from_end: usize, length: usize) {
-        let current = self.filled;
-
-        let start = current.checked_sub(offset_from_end).expect("in bounds");
-        let end = start.checked_add(length).expect("in bounds");
-
-        // Note also that the referenced string may overlap the current
-        // position; for example, if the last 2 bytes decoded have values
-        // X and Y, a string reference with <length = 5, distance = 2>
-        // adds X,Y,X,Y,X to the output stream.
-
-        if end > current {
-            if offset_from_end == 1 {
-                // this will just repeat this value many times
-                let element = self.buf[current - 1];
-                self.buf[current..][..length].fill(element);
-            } else {
-                for i in 0..length {
-                    self.buf[current + i] = self.buf[start + i];
-                }
-            }
-        } else {
-            Self::copy_chunked_within::<C>(self.buf, current, start, end)
-        }
-
-        // safety: we just copied length initialized bytes right beyond self.filled
-        unsafe { self.assume_init(length) };
-
-        self.advance(length);
-    }
-
-    #[inline(always)]
-    fn copy_chunked_within<C: Chunk>(
-        buf: &mut [MaybeUninit<u8>],
-        current: usize,
-        start: usize,
-        end: usize,
-    ) {
-        if (end - start).next_multiple_of(core::mem::size_of::<C>()) <= (buf.len() - current) {
-            unsafe {
-                Self::copy_chunk_unchecked::<C>(
-                    buf.as_ptr().add(start),
-                    buf.as_mut_ptr().add(current),
-                    buf.as_ptr().add(end),
-                )
-            }
-        } else {
-            // a full simd copy does not fit in the output buffer
-            buf.copy_within(start..end, current);
-        }
-    }
-
-    /// # Safety
-    ///
-    /// `src` must be safe to perform unaligned reads in `core::mem::size_of::<C>()` chunks until
-    /// `end` is reached. `dst` must be safe to (unalingned) write that number of chunks.
-    #[inline(always)]
-    unsafe fn copy_chunk_unchecked<C: Chunk>(
-        mut src: *const MaybeUninit<u8>,
-        mut dst: *mut MaybeUninit<u8>,
-        end: *const MaybeUninit<u8>,
-    ) {
-        while src < end {
-            let chunk = C::load_chunk(src);
-            C::store_chunk(dst, chunk);
-
-            src = src.add(core::mem::size_of::<C>());
-            dst = dst.add(core::mem::size_of::<C>());
-        }
-    }
-
     pub(crate) fn new_in(alloc: &Allocator<'a>, len: usize) -> Option<Self> {
         let buf = alloc.allocate_slice::<u8>(len)?;
 
@@ -338,73 +246,4 @@ unsafe fn slice_to_uninit_mut(slice: &mut [u8]) -> &mut [MaybeUninit<u8>] {
 // TODO: This could use `MaybeUninit::slice_assume_init` when it is stable.
 unsafe fn slice_assume_init(slice: &[MaybeUninit<u8>]) -> &[u8] {
     &*(slice as *const [MaybeUninit<u8>] as *const [u8])
-}
-
-trait Chunk {
-    /// Safety: must be valid to read a `Self::Chunk` value from `from` with an unaligned read.
-    unsafe fn load_chunk(from: *const MaybeUninit<u8>) -> Self;
-
-    /// Safety: must be valid to write a `Self::Chunk` value to `out` with an unaligned write.
-    unsafe fn store_chunk(out: *mut MaybeUninit<u8>, chunk: Self);
-}
-
-impl Chunk for u64 {
-    unsafe fn load_chunk(from: *const MaybeUninit<u8>) -> Self {
-        u64::to_le(core::ptr::read_unaligned(from.cast()))
-    }
-
-    unsafe fn store_chunk(out: *mut MaybeUninit<u8>, chunk: Self) {
-        core::ptr::copy_nonoverlapping(
-            chunk.to_le_bytes().as_ptr().cast(),
-            out,
-            core::mem::size_of::<Self>(),
-        )
-    }
-}
-
-#[cfg(target_arch = "x86_64")]
-impl Chunk for core::arch::x86_64::__m128i {
-    #[inline(always)]
-    unsafe fn load_chunk(from: *const MaybeUninit<u8>) -> Self {
-        core::arch::x86_64::_mm_loadu_si128(from.cast())
-    }
-
-    #[inline(always)]
-    unsafe fn store_chunk(out: *mut MaybeUninit<u8>, chunk: Self) {
-        core::arch::x86_64::_mm_storeu_si128(out as *mut Self, chunk);
-    }
-}
-
-#[cfg(target_arch = "x86_64")]
-impl Chunk for core::arch::x86_64::__m256i {
-    #[inline(always)]
-    unsafe fn load_chunk(from: *const MaybeUninit<u8>) -> Self {
-        core::arch::x86_64::_mm256_loadu_si256(from.cast())
-    }
-
-    #[inline(always)]
-    unsafe fn store_chunk(out: *mut MaybeUninit<u8>, chunk: Self) {
-        core::arch::x86_64::_mm256_storeu_si256(out as *mut Self, chunk);
-    }
-}
-
-#[cfg(target_arch = "x86_64")]
-impl Chunk for core::arch::x86_64::__m512i {
-    #[inline(always)]
-    unsafe fn load_chunk(from: *const MaybeUninit<u8>) -> Self {
-        // TODO AVX-512 is effectively unstable.
-        // We cross our fingers that LLVM optimizes this into a vmovdqu32
-        //
-        // https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html#text=_mm512_loadu_si512&expand=3420&ig_expand=4110
-        core::ptr::read_unaligned(from.cast())
-    }
-
-    #[inline(always)]
-    unsafe fn store_chunk(out: *mut MaybeUninit<u8>, chunk: Self) {
-        // TODO AVX-512 is effectively unstable.
-        // We cross our fingers that LLVM optimizes this into a vmovdqu32
-        //
-        // https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html#text=_mm512_storeu_si512&expand=3420&ig_expand=4110,6550
-        core::ptr::write_unaligned(out.cast(), chunk)
-    }
 }
